requirements.txt
pandas==2.1.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.9
gspread==5.9.0
gspread-dataframe==3.3.0